% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tuning_parameter.R
\name{calculate_tuning_parameter}
\alias{calculate_tuning_parameter}
\title{Calculate the tuning parameter for nuclear norm regularied estimators}
\usage{
calculate_tuning_parameter(
  data_frame,
  func,
  delta,
  R_max = 5,
  s = 10,
  iter_max = 10000,
  tol = 1e-08
)
}
\arguments{
\item{data_frame}{Panel data in long format (could be a data frame or 
matrix). The first column must contain individual IDs, the second column is
time, and the third column is the  outcome variable \(Y\); all remaining 
columns are treated as covariates in the regression model. The panel is 
required to be balanced.}

\item{func}{A character string indicating the model:
\code{"logit"},  \code{"probit"}, or \code{"poisson"}.}

\item{delta}{Positive scalar inflation factor used to set
\eqn{\varphi = (1+\delta)\sigma_1}, where \eqn{\sigma_1} is the leading
singular value of the residual matrix.}

\item{R_max}{Maximum number of factors allowed when determining the
number of factors.}

\item{s}{Initial step size  for both the NNR and MLE steps. 
This value serves as the starting step size for the line-search procedure.}

\item{iter_max}{Maximum number of iterations for optimizations.}

\item{tol}{Convergence tolerance passed to the optimization routines.}
}
\value{
A single numeric value giving the \eqn{\varphi}.
}
\description{
This function provides a data-driven way to calculate tuning parameter \eqn{\varphi} 
for non-linear nuclear-norm regularized (NNR) estimator with interactive fixed
effects. 
It supports both logistic and Poisson models.
}
\details{
The function first estimates a two–way fixed-effects model, obtains the 
fitted values, and sets the initial \eqn{\varphi} to be the  (1 + \eqn{\alpha}) * leading 
singular value of the residual matrix. It then runs the NNR step, 
selects the number of factors, refines the estimates via MLE, and 
updates \eqn{\varphi} based on the new residuals.
}
\examples{
\dontrun{
data_list <- gen_data_poisson(N = 200, T = 200, num_factor = 2, num_Z = 1, beta = c(0.2))
data_frame <- list_to_data_frame(data_list) 
phi <- calculate_tuning_parameter(
  data_frame = data_frame,
  func       = "poisson",
  delta      = 0.05,
  R_max      = 5
)
}

}
