% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_num_factor.R
\name{determine_num_factor}
\alias{determine_num_factor}
\title{Determine the Number of Factors using eigenvalue ratio test.}
\usage{
determine_num_factor(sigma, R_max)
}
\arguments{
\item{sigma}{A numeric vector of non-negative values, typically singular 
values (or eigenvalues) sorted in non-increasing order. Must have length 
at least \code{R_max + 1}.}

\item{R_max}{An integer indicating the maximum number of factors to consider.}
}
\value{
An integer giving the estimated number of factor.
}
\description{
Computes the number of factors using eigenvalue ratio test 
(see Ahn, S. C., & Horenstein, A. R. (2013))
}
\details{
The algorithm follows two steps:
\itemize{
  \item If any element \code{sigma[i] <= 1e-6}, the estimated number of 
  factors is set to \code{i - 1}.
  \item Otherwise, it identifies the index \code{i} that maximizes 
  \code{sigma[i] / sigma[i + 1]}, capturing a sharp drop in the first \code{R_max + 1} 
  elements of \code{sigma}. 
}
}
\examples{
sigma <- c(5.0, 3.1, 1.9, 0.2, 1e-8, 1e-9)
determine_num_factor(sigma, R_max = 4) 

}
\references{
Ahn, S. C., & Horenstein, A. R. (2013). Eigenvalue ratio test for the number of factors. Econometrica, 81(3), 1203-1227.
}
