% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pool_estimate.R
\name{Pool_estimate}
\alias{Pool_estimate}
\title{Pooled Estimator for Panel Data}
\usage{
Pool_estimate(data_frame, func)
}
\arguments{
\item{data_frame}{Panel data in long format (could be a data frame or 
matrix). The first column must contain individual IDs, the second column is
time, and the third column is the  outcome variable \(Y\); all remaining 
columns are treated as covariates in the regression model. The panel is 
required to be balanced.}

\item{func}{A character string indicating the model:
\code{"logit"},  \code{"probit"}, or \code{"poisson"}.}
}
\value{
A numeric vector of pooled coefficient estimates.
}
\description{
Computes the pooled estimator for a panel dataset by fitting a generalized 
linear model without fixed effects.
}
\details{
The function fits a pooled GLM without individual or time effects. For 
\code{func="poisson"}, it uses Poisson regression; for 
\code{func="logit"}, it uses logistic regression.
}
\examples{
\dontrun{
data_list <- gen_data_poisson(N = 100, T = 50, num_factor = 2, num_Z = 1, beta = c(0.2))
data_frame <- list_to_data_frame(data_list)

beta <- Pool_estimate(data_frame, "poisson")
}

}
